Add AboutMe component. Update about page.
