+++ components/about/AboutMe.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const AboutMe = ({ title, description, fadeClassName }) => (
+  <div className={`aboutme ${fadeClassName()}`}>
+    <div className={`header`}>{title}</div>
+    <div className={`description`}>
+      <p>{description}</p>
+    </div>
+  </div>
+);
+
+export default AboutMe;
